extract RowToggleIcon component
