+++ frontend/src/metabase/visualizations/visualizations/PivotTable/PivotTable.jsx
@@ -3,7 +3,7 @@ import React, { Component } from "react";
 import { t } from "ttag";
 import cx from "classnames";
 import _ from "underscore";
-import { getIn, updateIn } from "icepick";
+import { getIn } from "icepick";
 import { Grid, Collection, ScrollSync, AutoSizer } from "react-virtualized";
 
 import { findDOMNode } from "react-dom";
@@ -12,7 +12,6 @@ import { getScrollBarSize } from "metabase/lib/dom";
 import ChartSettingsTableFormatting from "metabase/visualizations/components/settings/ChartSettingsTableFormatting";
 
 import Ellipsified from "metabase/core/components/Ellipsified";
-import Icon from "metabase/components/Icon";
 import {
   COLLAPSED_ROWS_SETTING,
   COLUMN_SPLIT_SETTING,
@@ -30,11 +29,13 @@ import { ChartSettingIconRadio } from "metabase/visualizations/components/settin
 
 import { PLUGIN_SELECTORS } from "metabase/plugins";
 import { isDimension } from "metabase-lib/types/utils/isa";
+
+import { RowToggleIcon } from "./RowToggleIcon";
+
 import {
   PivotTableRoot,
   PivotTableCell,
   PivotTableTopLeftCellsContainer,
-  RowToggleIconRoot,
   CELL_HEIGHT,
 } from "./PivotTable.styled";
 
@@ -591,77 +592,6 @@ class PivotTable extends Component {
 }
 
 export default connect(mapStateToProps)(PivotTable);
-
-function RowToggleIcon({
-  value,
-  settings,
-  updateSettings,
-  hideUnlessCollapsed,
-  rowIndex,
-  hasCustomColors,
-  isNightMode,
-}) {
-  if (value == null) {
-    return null;
-  }
-  const setting = settings[COLLAPSED_ROWS_SETTING];
-  const ref = JSON.stringify(value);
-  const isColumn = !Array.isArray(value);
-  const columnRef = isColumn ? null : JSON.stringify(value.length);
-  const settingValue = setting.value || [];
-  const isColumnCollapsed = !isColumn && settingValue.includes(columnRef);
-  const isCollapsed = settingValue.includes(ref) || isColumnCollapsed;
-  if (hideUnlessCollapsed && !isCollapsed) {
-    // subtotal rows shouldn't have an icon unless the section is collapsed
-    return null;
-  }
-
-  // The giant nested ternary below picks the right function to toggle the current button.
-  // That depends on whether we're a row or column header and whether we're open or closed.
-  const toggle =
-    isColumn && !isCollapsed // click on open column
-      ? settingValue =>
-          settingValue
-            .filter(v => {
-              const parsed = JSON.parse(v);
-              return !(Array.isArray(parsed) && parsed.length === value);
-            }) // remove any already collapsed items in this column
-            .concat(ref) // add column to list
-      : !isColumn && isColumnCollapsed // single row in collapsed column
-      ? settingValue =>
-          settingValue
-            .filter(v => v !== columnRef) // remove column from list
-            .concat(
-              // add other rows in this columns so they stay closed
-              rowIndex
-                .filter(
-                  item =>
-                    // equal length means they're in the same column
-                    item.length === value.length &&
-                    // but not exactly this item
-                    !_.isEqual(item, value),
-                )
-                // serialize those paths
-                .map(item => JSON.stringify(item)),
-            )
-      : isCollapsed // closed row or column
-      ? settingValue => settingValue.filter(v => v !== ref)
-      : // open row or column
-        settingValue => settingValue.concat(ref);
-
-  return (
-    <RowToggleIconRoot
-      onClick={e => {
-        e.stopPropagation();
-        updateSettings({
-          [COLLAPSED_ROWS_SETTING]: updateIn(setting, ["value"], toggle),
-        });
-      }}
-    >
-      <Icon name={isCollapsed ? "add" : "dash"} size={8} />
-    </RowToggleIconRoot>
-  );
-}
 
 function Cell({
   value,
